bio_t5/requirements.txt
accelerate==0.31.0
datasets == 2.13.1
sentencepiece != 0.1.92
transformers==4.33.3
neptune
pdbpp
notebook
protobuf==3.20.*
pyyaml
pynvml
hydra-core==1.3.2
evaluate==0.4.2
nltk
absl-py
rouge_score
deepchem
selfies
tabulate
schedule
scikit-learn
fsspec==2023.6.0
rdkit==2022.09.3